Move config_entry type to coordinator file
