+++ homeassistant/components/energyzero/__init__.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
@@ -10,13 +9,11 @@ from homeassistant.helpers import config_validation as cv
 from homeassistant.helpers.typing import ConfigType
 
 from .const import DOMAIN
-from .coordinator import EnergyZeroDataUpdateCoordinator
+from .coordinator import EnergyZeroConfigEntry, EnergyZeroDataUpdateCoordinator
 from .services import async_setup_services
 
 PLATFORMS: list[Platform] = [Platform.SENSOR]
 CONFIG_SCHEMA = cv.config_entry_only_config_schema(DOMAIN)
-
-type EnergyZeroConfigEntry = ConfigEntry[EnergyZeroDataUpdateCoordinator]
 
 
 async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
@@ -30,7 +27,7 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
 async def async_setup_entry(hass: HomeAssistant, entry: EnergyZeroConfigEntry) -> bool:
     """Set up EnergyZero from a config entry."""
 
-    coordinator = EnergyZeroDataUpdateCoordinator(hass)
+    coordinator = EnergyZeroDataUpdateCoordinator(hass, entry)
     try:
         await coordinator.async_config_entry_first_refresh()
     except ConfigEntryNotReady:
